refactor: flatten WebUI crate layout
