+++ .github/dependabot.yml
@@ -1,7 +1,7 @@
 version: 2
 updates:
   - package-ecosystem: cargo
-    directory: /webui
+    directory: /
     schedule:
       interval: monthly
     open-pull-requests-limit: 1
