+++ .github/workflows/webui-release.yml
@@ -39,14 +39,12 @@ jobs:
 
       - name: Cache Cargo
         uses: Swatinem/rust-cache@v2
-        with:
-          workspaces: webui -> ../target
 
       - name: Test
-        run: cargo test --manifest-path webui/Cargo.toml --target-dir target
+        run: cargo test --target-dir target
 
       - name: Build release
-        run: cargo build --release --manifest-path webui/Cargo.toml --target-dir target --target ${{ matrix.target }}
+        run: cargo build --release --target-dir target --target ${{ matrix.target }}
 
       - name: Verify release version
         run: |
@@ -104,14 +102,12 @@ jobs:
 
       - name: Cache Cargo
         uses: Swatinem/rust-cache@v2
-        with:
-          workspaces: webui -> ../target
 
       - name: Test
-        run: cargo test --manifest-path webui/Cargo.toml --target-dir target
+        run: cargo test --target-dir target
 
       - name: Build release
-        run: cargo build --release --manifest-path webui/Cargo.toml --target-dir target --target x86_64-unknown-linux-gnu
+        run: cargo build --release --target-dir target --target x86_64-unknown-linux-gnu
 
       - name: Verify release version
         run: |
