+++ Cargo.toml
@@ -1,8 +1,19 @@
-[workspace]
-members = ["webui"]
-resolver = "2"
-
-[workspace.package]
+[package]
+name = "herdr-webui"
+version = "0.0.15"
 edition = "2021"
 license = "AGPL-3.0-or-later"
 repository = "https://github.com/alecuba16/herdr-webui"
+description = "Web UI for an existing Herdr server"
+
+[dependencies]
+axum = { version = "0.8", features = ["ws"] }
+bincode = { version = "2", features = ["serde"] }
+interprocess = "2.4.2"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.11"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "time", "net"] }
+
+[dev-dependencies]
+tower = { version = "0.5", features = ["util"] }
