+++ src/assets/app.js
