+++ src/assets/app_core.js
