+++ src/assets/app_core.test.mjs
