+++ src/assets/mobile_attention.js
