+++ src/assets/mobile_core.js
