+++ src/assets/mobile_load.test.mjs
