+++ src/assets/mobile_settings.js
