+++ src/assets/mobile_terminal.js
