+++ src/protocol.rs
@@ -1,0 +1,256 @@
+use std::io::{Read, Write};
+
+use serde::{Deserialize, Serialize};
+
+pub(crate) fn write_message<W: Write, M: Serialize>(writer: &mut W, msg: &M) -> Result<(), String> {
+    let payload = bincode::serde::encode_to_vec(msg, bincode::config::standard())
+        .map_err(|err| err.to_string())?;
+    let len = u32::try_from(payload.len()).map_err(|_| "payload too large".to_string())?;
+    writer
+        .write_all(&len.to_le_bytes())
+        .map_err(|err| err.to_string())?;
+    writer.write_all(&payload).map_err(|err| err.to_string())?;
+    writer.flush().map_err(|err| err.to_string())
+}
+
+pub(crate) fn read_message<R: Read, M: for<'de> Deserialize<'de>>(
+    reader: &mut R,
+    max_frame_size: usize,
+) -> Result<M, String> {
+    let mut len_buf = [0u8; 4];
+    reader
+        .read_exact(&mut len_buf)
+        .map_err(|err| err.to_string())?;
+    let len = u32::from_le_bytes(len_buf) as usize;
+    if len > max_frame_size {
+        return Err(format!("frame size {len} exceeds maximum {max_frame_size}"));
+    }
+    let mut payload = vec![0u8; len];
+    reader
+        .read_exact(&mut payload)
+        .map_err(|err| err.to_string())?;
+    let (msg, consumed) = bincode::serde::decode_from_slice(&payload, bincode::config::standard())
+        .map_err(|err| err.to_string())?;
+    if consumed != len {
+        return Err("trailing bytes in frame".into());
+    }
+    Ok(msg)
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum RenderEncoding {
+    SemanticFrame,
+    TerminalAnsi,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientKeybindings {
+    Server,
+    Local { keys_toml: String },
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientLaunchMode {
+    App,
+    TerminalAttach,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientMessage {
+    Hello {
+        version: u32,
+        cols: u16,
+        rows: u16,
+        cell_width_px: u32,
+        cell_height_px: u32,
+        requested_encoding: RenderEncoding,
+        keybindings: ClientKeybindings,
+        launch_mode: ClientLaunchMode,
+    },
+    Input {
+        data: Vec<u8>,
+    },
+    ClipboardImage {
+        extension: String,
+        data: Vec<u8>,
+    },
+    Resize {
+        cols: u16,
+        rows: u16,
+        cell_width_px: u32,
+        cell_height_px: u32,
+    },
+    Detach,
+    AttachTerminal {
+        terminal_id: String,
+        takeover: bool,
+    },
+    AttachScroll {
+        source: AttachScrollSource,
+        direction: AttachScrollDirection,
+        lines: u16,
+        column: Option<u16>,
+        row: Option<u16>,
+        modifiers: u8,
+    },
+    InputEvents {
+        events: Vec<ClientInputEvent>,
+    },
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum AttachScrollDirection {
+    Up,
+    Down,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum AttachScrollSource {
+    Wheel,
+    PageKey { input: Vec<u8> },
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientInputEvent {
+    Key {
+        code: ClientKeyCode,
+        modifiers: u8,
+        kind: ClientKeyKind,
+    },
+    Mouse {
+        kind: ClientMouseKind,
+        column: u16,
+        row: u16,
+        modifiers: u8,
+    },
+    Paste {
+        text: String,
+    },
+    FocusGained,
+    FocusLost,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientKeyKind {
+    Press,
+    Repeat,
+    Release,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientKeyCode {
+    Backspace,
+    Enter,
+    Left,
+    Right,
+    Up,
+    Down,
+    Home,
+    End,
+    PageUp,
+    PageDown,
+    Tab,
+    BackTab,
+    Delete,
+    Insert,
+    Esc,
+    Char(char),
+    F(u8),
+    Null,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientMouseButton {
+    Left,
+    Right,
+    Middle,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ClientMouseKind {
+    Down(ClientMouseButton),
+    Up(ClientMouseButton),
+    Drag(ClientMouseButton),
+    Moved,
+    ScrollUp,
+    ScrollDown,
+    ScrollLeft,
+    ScrollRight,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) struct CellData {
+    symbol: String,
+    fg: u32,
+    bg: u32,
+    modifier: u16,
+    skip: bool,
+    hyperlink: Option<u32>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) struct CursorState {
+    x: u16,
+    y: u16,
+    visible: bool,
+    #[serde(default)]
+    shape: u8,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) struct FrameData {
+    cells: Vec<CellData>,
+    width: u16,
+    height: u16,
+    cursor: Option<CursorState>,
+    hyperlinks: Vec<String>,
+    graphics: Vec<u8>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) struct TerminalFrame {
+    pub(crate) seq: u64,
+    pub(crate) width: u16,
+    pub(crate) height: u16,
+    pub(crate) full: bool,
+    pub(crate) bytes: Vec<u8>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum NotifyKind {
+    Sound,
+    Toast,
+    SystemToast,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub(crate) enum ServerMessage {
+    Welcome {
+        version: u32,
+        encoding: RenderEncoding,
+        error: Option<String>,
+    },
+    Frame(FrameData),
+    Terminal(TerminalFrame),
+    Graphics {
+        bytes: Vec<u8>,
+    },
+    ServerShutdown {
+        reason: Option<String>,
+    },
+    Notify {
+        kind: NotifyKind,
+        message: String,
+        body: Option<String>,
+    },
+    Clipboard {
+        data: String,
+    },
+    WindowTitle {
+        title: Option<String>,
+    },
+    ReloadSoundConfig,
+    MouseCapture {
+        enabled: bool,
+    },
+}
