--- webui/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "herdr-webui"
-version = "0.0.15"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Web UI for an existing Herdr server"
-
-[dependencies]
-axum = { version = "0.8", features = ["ws"] }
-bincode = { version = "2", features = ["serde"] }
-interprocess = "2.4.2"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-sha2 = "0.11"
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "time", "net"] }
-
-[dev-dependencies]
-tower = { version = "0.5", features = ["util"] }
